refactor de gyro
